+++ src/main/java/cn/taketoday/context/io/UrlBasedResource.java
@@ -93,4 +93,5 @@ public class UrlBasedResource extends AbstractResource {
         }
         return new UrlBasedResource(new URL(this.url, relativePath));
     }
+
 }