+++ src/test/java/test/context/ApplicationContextTest.java
@@ -31,13 +31,16 @@ import org.junit.Before;
 import org.junit.Test;
 
 import cn.taketoday.context.ApplicationContext;
+import cn.taketoday.context.ConfigurableApplicationContext;
 import cn.taketoday.context.StandardApplicationContext;
 import cn.taketoday.context.bean.BeanDefinition;
 import cn.taketoday.context.bean.PropertyValue;
 import cn.taketoday.context.env.ConfigurableEnvironment;
+import cn.taketoday.context.event.ContextStartedEvent;
 import cn.taketoday.context.exception.BeanDefinitionStoreException;
 import cn.taketoday.context.exception.NoSuchBeanDefinitionException;
 import cn.taketoday.context.factory.BeanDefinitionRegistry;
+import cn.taketoday.context.listener.ApplicationListener;
 import lombok.extern.slf4j.Slf4j;
 import test.demo.config.Config;
 import test.demo.config.ConfigFactoryBean;
@@ -246,19 +249,41 @@ public class ApplicationContextTest {
         }
     }
 
+//    @Test
+//    public void test_OnConstructor() throws NoSuchBeanDefinitionException, BeanDefinitionStoreException {
+//
+//        try (ApplicationContext applicationContext = new StandardApplicationContext()) {
+//            applicationContext.loadContext("");
+//            ConfigurableEnvironment environment = applicationContext.getEnvironment();
+//            BeanDefinitionRegistry beanDefinitionRegistry = environment.getBeanDefinitionRegistry();
+//            System.err.println(beanDefinitionRegistry.getBeanDefinitionsMap());
+//
+//            Config bean = applicationContext.getBean(Config.class);
+//            System.out.println(bean);
+//            assert bean != null;
+//            assert bean.getUser() != null;
+//        }
+//    }
+
+    boolean i = false;
+
     @Test
-    public void test_OnConstructor() throws NoSuchBeanDefinitionException, BeanDefinitionStoreException {
+    public void testAddApplicationListener() throws NoSuchBeanDefinitionException, BeanDefinitionStoreException {
 
-        try (ApplicationContext applicationContext = new StandardApplicationContext()) {
+        try (ConfigurableApplicationContext applicationContext = new StandardApplicationContext()) {
+
+            applicationContext.addApplicationListener(new ApplicationListener<ContextStartedEvent>() {
+
+                @Override
+                public void onApplicationEvent(ContextStartedEvent event) {
+                    i = true;
+                    System.err.println(i);
+                }
+            });
+
             applicationContext.loadContext("");
-            ConfigurableEnvironment environment = applicationContext.getEnvironment();
-            BeanDefinitionRegistry beanDefinitionRegistry = environment.getBeanDefinitionRegistry();
-            System.err.println(beanDefinitionRegistry.getBeanDefinitionsMap());
 
-            Config bean = applicationContext.getBean(Config.class);
-            System.out.println(bean);
-            assert bean != null;
-            assert bean.getUser() != null;
+            assert i;
         }
     }